+++ src/dataDisplay/tableBody.jsx
@@ -1,37 +1,59 @@
-'use strict';
-
-import React, { Component } from 'react';
 import ClassNames from 'classnames';
 import PropTypes from 'prop-types';
+import React from 'react';
 
-class TableBody extends Component {
-    render() {
+const propTypes = {
+    /**
+     * The content of the TableBody
+     */
+    children: PropTypes.node,
+    /**
+     * Assign additional class names to TableBody.
+     */
+    className: PropTypes.string,
+    /**
+     * The `id` of the TableBody.
+     */
+    id: PropTypes.string,
+    style: PropTypes.shape({}),
+};
+
+const defaultProps = {
+    children: null,
+    className: null,
+    id: null,
+    style: null,
+};
+
+const TableBody = React.forwardRef(
+    // eslint-disable-next-line prefer-arrow-callback
+    function TableBody(props, ref) {
         const {
+            children,
             className,
             id,
             style,
-        } = this.props;
+        } = props;
+
         const containerClasses = ClassNames(
             'table-body',
-            className
+            className,
         );
 
         return (
             <tbody
                 className={containerClasses}
                 id={id}
+                ref={ref}
                 style={style}
             >
-                {this.props.children}
+                {children}
             </tbody>
         );
-    }
-}
+    },
+);
 
-TableBody.propTypes = {
-    className: PropTypes.string,
-    id: PropTypes.string,
-    style: PropTypes.shape({}),
-};
+TableBody.propTypes = propTypes;
+TableBody.defaultProps = defaultProps;
 
 export default TableBody;