+++ src/templates/dataGridTable.jsx
@@ -1,9 +1,17 @@
-import _ from 'lodash';
+import {
+    flatMap,
+    isFunction,
+    map,
+    isArray,
+    isNil,
+} from 'lodash';
 import ClassNames from 'classnames';
 import PropTypes from 'prop-types';
 import React from 'react';
+import { SORTABLE_PROP_TYPES } from './dataGridConstants';
 import DataGridTableRow from './dataGridTableRow';
 import Table from '../dataDisplay/table';
+import DataGridTableReactSortable from './dataGridTableReactSortable';
 
 const propTypes = {
     bleed: PropTypes.bool,
@@ -14,10 +22,14 @@ const propTypes = {
         id: PropTypes.string,
         style: PropTypes.shape({}),
     })).isRequired,
-    data: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
+    data: PropTypes.arrayOf(
+        PropTypes.oneOfType([
+            PropTypes.arrayOf(PropTypes.shape({})),
+            PropTypes.shape({}),
+        ]),
+    ).isRequired,
     dropShadow: PropTypes.bool,
     fontSize: PropTypes.string,
-    handle: PropTypes.bool,
     id: PropTypes.string.isRequired,
     idPrefix: PropTypes.string,
     resizableColumnWidthPercentage: PropTypes.number,
@@ -26,6 +38,7 @@ const propTypes = {
         'small',
         'medium',
     ]),
+    sortable: PropTypes.arrayOf(SORTABLE_PROP_TYPES),
     stickyColumns: PropTypes.number,
     stretch: PropTypes.oneOfType([
         PropTypes.oneOf(['very']),
@@ -39,45 +52,146 @@ const defaultProps = {
     className: undefined,
     dropShadow: false,
     fontSize: 'xsmall',
-    handle: false,
     idPrefix: 'base',
     resizableColumnWidthPercentage: undefined,
-    rowProps: undefined,
+    rowProps: () => ({
+        className: null,
+        id: null,
+        onClick: null,
+        selected: null,
+        style: null,
+    }),
     size: 'small',
+    sortable: null,
     stickyColumns: undefined,
     stretch: false,
     style: undefined,
 };
 
 class DataGridTable extends React.PureComponent {
+    constructor(props) {
+        super(props);
+
+        this.onSortableChange = this.onSortableChange.bind(this);
+    }
+
+    onSortableChange(newData, index) {
+        const {
+            sortable,
+        } = this.props;
+
+        if (isFunction(sortable[index].onChange)) {
+            sortable[index].onChange(newData, index);
+        }
+    }
+
     render() {
         const {
             classNamePrefix,
             bleed: bleedProp,
             className,
             columns,
-            data,
+            data: dataProp,
             dropShadow,
             fontSize,
-            handle,
             id,
             idPrefix,
             rowProps,
             resizableColumnWidthPercentage,
             size,
+            sortable,
             stickyColumns,
             stretch,
             style,
         } = this.props;
-        const bleed = bleedProp ? 'very' : stretch;
-        const containerClasses = ClassNames('ui', `${classNamePrefix}_table`, className);
+
+        const isSelectable = isFunction(rowProps().onClick);
+
+        let rowKeyNum = 1;
+
+        const tableRows = (data, handle) => flatMap(data, (row) => {
+            rowKeyNum += 1;
+
+            return (
+                <DataGridTableRow
+                    classNamePrefix={classNamePrefix}
+                    columns={columns}
+                    handle={handle}
+                    id={id}
+                    idPrefix={idPrefix}
+                    isClickable={isSelectable}
+                    key={`tableBodyRow-${row.id || rowKeyNum}`}
+                    row={row}
+                    rowIndex={rowKeyNum}
+                    rowProps={rowProps(row)}
+                />
+            );
+        });
+
+        let tableBody;
         const bodyClasses = ClassNames({ [`${classNamePrefix}_drop_shadow`]: dropShadow });
-        const isSelectable =
-            !_.isUndefined(rowProps) && _.isFunction(rowProps().onClick);
+
+        if (sortable === true) {
+            tableBody = (
+                <DataGridTableReactSortable
+                    arrayIndex={0}
+                    className={bodyClasses}
+                    list={dataProp}
+                    sortable={{
+                        ...sortable,
+                        filter: `${classNamePrefix}-filter`,
+                    }}
+                >
+                    {tableRows(
+                        dataProp,
+                        !isNil(sortable[0].handle) ? sortable[0].handle : true,
+                    )}
+                </DataGridTableReactSortable>
+            );
+        } else if (isArray(sortable) && isArray(dataProp[0])) {
+            let tBodyKeyNum = 1;
+
+            tableBody = map(dataProp, (arrayOfData, arrayIndex) => {
+                const hasHandle = !isNil(sortable[arrayIndex].handle) ?
+                    sortable[arrayIndex].handle :
+                    true;
+
+                tBodyKeyNum += 1;
+
+                return (
+                    <DataGridTableReactSortable
+                        arrayIndex={arrayIndex}
+                        className={bodyClasses}
+                        key={`tbody--${tBodyKeyNum}`}
+                        list={arrayOfData}
+                        sortable={{
+                            ...sortable,
+                            filter: `${classNamePrefix}-filter`,
+                        }}
+                    >
+                        {tableRows(
+                            arrayOfData,
+                            hasHandle,
+                        )}
+                    </DataGridTableReactSortable>
+                );
+            });
+        } else {
+            tableBody = (
+                <Table.Body
+                    className={bodyClasses}
+                >
+                    {tableRows(dataProp)}
+                </Table.Body>
+            );
+        }
+
+        const rootClasses = ClassNames('ui', `${classNamePrefix}_table`, className);
+        const bleed = bleedProp ? 'very' : stretch;
 
         return (
             <div
-                className={containerClasses}
+                className={rootClasses}
                 style={style}
             >
                 <Table
@@ -92,12 +206,14 @@ class DataGridTable extends React.PureComponent {
                 >
                     <Table.Header>
                         <Table.Row>
-                            {_.map(columns, (column, index) => {
+                            {map(columns, (column, index) => {
                                 const headerCellClasses = ClassNames(
                                     `${classNamePrefix}_table_header_cell`,
                                     column.className,
                                 );
+
                                 const cellId = column.id || `${classNamePrefix}_table_${id}_header_${idPrefix}-${index}`;
+
                                 return (
                                     <Table.HeaderCell
                                         className={headerCellClasses}
@@ -112,22 +228,7 @@ class DataGridTable extends React.PureComponent {
                         </Table.Row>
                     </Table.Header>
 
-                    <Table.Body className={bodyClasses}>
-                        {_.map(data, (row, index) => (
-                            <DataGridTableRow
-                                classNamePrefix={classNamePrefix}
-                                columns={columns}
-                                handle={handle}
-                                id={id}
-                                idPrefix={idPrefix}
-                                isClickable={isSelectable}
-                                key={`tableBodyRow-${row.id || index}`}
-                                row={row}
-                                rowIndex={index}
-                                rowProps={rowProps(row)}
-                            />
-                        ))}
-                    </Table.Body>
+                    {tableBody}
                 </Table>
             </div>
         );